Fix a few warnings in PrintHostDialogs, UploadDialogs

+++ src/slic3r/GUI/PrintHostDialogs.cpp
@@ -250,7 +250,7 @@ void PrintHostQueueDialog::on_list_select()
 
 void PrintHostQueueDialog::on_progress(Event &evt)
 {
-    wxCHECK_RET(evt.job_id < job_list->GetItemCount(), "Out of bounds access to job list");
+    wxCHECK_RET(evt.job_id < (size_t)job_list->GetItemCount(), "Out of bounds access to job list");
 
     if (evt.progress < 100) {
         set_state(evt.job_id, ST_PROGRESS);
@@ -265,7 +265,7 @@ void PrintHostQueueDialog::on_progress(Event &evt)
 
 void PrintHostQueueDialog::on_error(Event &evt)
 {
-    wxCHECK_RET(evt.job_id < job_list->GetItemCount(), "Out of bounds access to job list");
+    wxCHECK_RET(evt.job_id < (size_t)job_list->GetItemCount(), "Out of bounds access to job list");
 
     set_state(evt.job_id, ST_ERROR);
 
@@ -280,7 +280,7 @@ void PrintHostQueueDialog::on_error(Event &evt)
 
 void PrintHostQueueDialog::on_cancel(Event &evt)
 {
-    wxCHECK_RET(evt.job_id < job_list->GetItemCount(), "Out of bounds access to job list");
+    wxCHECK_RET(evt.job_id < (size_t)job_list->GetItemCount(), "Out of bounds access to job list");
 
     set_state(evt.job_id, ST_CANCELLED);
     job_list->SetValue(wxVariant(0), evt.job_id, COL_PROGRESS);
